+++ R/regroupReads.R
@@ -42,12 +42,12 @@
 #'
 #' # regroup by read annotation (variant label)
 #' # ... across samples
-#' sere <- regroupReadsByColData(se, colNames = "variant_label",
+#' sere <- regroupReadsByColData(se, colNames = "readInfo:variant_label",
 #'                               withinSample = FALSE)
 #' sere
 #' lapply(assay(sere, "mod_prob"), ncol)
 #' # ... within sample
-#' sere <- regroupReadsByColData(se, colNames = "variant_label",
+#' sere <- regroupReadsByColData(se, colNames = "readInfo:variant_label",
 #'                               withinSample = TRUE)
 #' sere
 #' lapply(assay(sere, "mod_prob"), ncol)
@@ -82,14 +82,22 @@ regroupReads <- function(se, readGroups) {
     # exclude groups without reads
     readGroups <- readGroups[lengths(readGroups) > 0]
 
-    # check that the modbase is consistent for each read group
-    mbmap <- setNames(
-        rep(se$modbase, vapply(assay(se, rlAssays[1]), ncol, 0L)),
-        unlist(lapply(assay(se, rlAssays[1]), colnames))
-    )
-    modbase <- lapply(readGroups, function(rg) unique(mbmap[rg]))
-    if (any(lengths(modbase) > 1)) {
-        cli_abort("Some read groups correspond to reads with different modbases")
+    # check that the modbase is consistent for each read group if it exists
+    if (!is.null(se$modbase)) {
+        mbmap <- setNames(
+            rep(se$modbase, vapply(assay(se, rlAssays[1]), ncol, 0L)),
+            unlist(lapply(assay(se, rlAssays[1]), colnames))
+        )
+        modbase <- lapply(readGroups, function(rg) unique(mbmap[rg]))
+        if (any(lengths(modbase) > 1)) {
+            cli_abort("Some read groups correspond to reads with different modbases")
+        }
+        cdata <- DataFrame(sample = names(readGroups),
+                           modbase = unlist(modbase),
+                           n_reads = lengths(readGroups))
+    } else {
+        cdata <- DataFrame(sample = names(readGroups),
+                           n_reads = lengths(readGroups))
     }
 
     # generate regrouped assays
@@ -106,19 +114,17 @@ regroupReads <- function(se, readGroups) {
     })
 
     # generate regrouped colData columns
-    cdata <- DataFrame(sample = names(readGroups),
-                       modbase = unlist(modbase),
-                       n_reads = lengths(readGroups))
     for (rlc in rlCols) {
         if (is(colData(se)[[rlc]][[1]], "data.frame") ||
             is(colData(se)[[rlc]][[1]], "DataFrame")) {
             tmp <- do.call(rbind, colData(se)[[rlc]])
-            cdata[[rlc]] <- S4Vectors::SimpleList(lapply(setNames(names(readGroups), names(readGroups)),
-                                   function(nm) {
-                                       tmp2 <- tmp[readGroups[[nm]], , drop = FALSE]
-                                       rownames(tmp2) <- paste0(nm, "-", rownames(tmp2))
-                                       tmp2
-                                   }))
+            cdata[[rlc]] <- S4Vectors::SimpleList(
+                lapply(setNames(names(readGroups), names(readGroups)),
+                       function(nm) {
+                           tmp2 <- tmp[readGroups[[nm]], , drop = FALSE]
+                           rownames(tmp2) <- paste0(nm, "-", rownames(tmp2))
+                           tmp2
+                       }))
         } else if (is(colData(se)[[rlc]][[1]], "IRangesList")) {
             irl <- do.call(c, unname(colData(se)[[rlc]]))
             cdata[[rlc]] <- lapply(setNames(names(readGroups), names(readGroups)),
@@ -148,9 +154,12 @@ regroupReads <- function(se, readGroups) {
 #' @param colNames A character vector corresponding to the names of annotation
 #'     (\code{colData}) columns, the combination of which represent the
 #'     desired grouping of the reads. The names can be either columns of
-#'     \code{colData(se)$readInfo}, or columns in \code{colData(se)} itself.
-#'     If a column name is present in both of these, the column in
-#'     \code{colData(se)$readInfo} will be used.
+#'     \code{colData(se)} itself, or columns in a nested (read-level)
+#'     annotation column of \code{colData(se)}. In the latter case, the
+#'     \code{colNames} should be of the form \code{outerColName:innerColName},
+#'     where \code{outerColName} is a column name in \code{colData(se)},
+#'     and \code{innerColName} is a column name in each element of
+#'     \code{colData(se)[[outerColName]]}.
 #' @param withinSample A logical scalar, indicating whether the regrouping
 #'     should be done within each current sample (column of \code{se}) or not.
 #'     If \code{FALSE} (default), reads are pooled across samples before
@@ -162,29 +171,63 @@ regroupReads <- function(se, readGroups) {
 #' @importFrom SummarizedExperiment colData
 #' @importFrom cli cli_abort
 regroupReadsByColData <- function(se, colNames, withinSample = FALSE) {
-    .assertVector(x = colNames, type = "character",
-                  validValues = c(colnames(colData(se)),
-                                  colnames(se$readInfo[[1]])))
+    checkSEValidity(se, verbose = FALSE)
+    rlAssays <- getReadLevelAssayNames(se)
+    if (length(rlAssays) == 0) {
+        cli_abort("{.arg se} does not contain any read-level assays")
+    }
+    .assertVector(x = colNames, type = "character")
+    colNamesTopLevel <- colNames[!grepl(":", colNames, fixed = TRUE)]
+    colNamesNested <- colNames[grepl(":", colNames, fixed = TRUE)]
+    matches <- regexec("^([^:]+):(.+)$", colNamesNested)
+    tmp <- regmatches(colNamesNested, matches)
+    colNamesNested <- setNames(vapply(tmp, "[", 3, FUN.VALUE = ""),
+                               vapply(tmp, "[", 2, FUN.VALUE = ""))
+    for (cn in colNamesTopLevel) {
+        .assertScalar(x = cn, type = "character", validValues = colnames(colData(se)))
+    }
+    for (i in seq_along(colNamesNested)) {
+        .assertScalar(x = names(colNamesNested)[i],
+                      type = "character", validValues = getReadLevelColDataNames(se))
+        .assertScalar(x = colNamesNested[i], type = "character",
+                      validValues = colnames(se[[names(colNamesNested)[i]]][[1]]))
+    }
     .assertScalar(x = withinSample, type = "logical")
 
-    readInfo <- cbind(sample = rep(se$sample, vapply(se$readInfo, nrow, 0L)),
-                      read_id = unlist(lapply(se$readInfo, rownames)),
-                      do.call(rbind, se$readInfo))
+    nReadsPerSample <- vapply(assay(se, rlAssays[1]), ncol, 0L)
+    readInfo <- data.frame(
+        sample = rep(se$sample, nReadsPerSample),
+        read_id = unlist(lapply(assay(se, rlAssays[1]), colnames))
+    )
+    for (i in seq_along(colNamesNested)) {
+        if (!is.atomic(se[[names(colNamesNested)[i]]][[1]][[colNamesNested[i]]])) {
+            cli_abort("The {.var {names(colNamesNested)[i]}:{colNamesNested[i]}} column is not atomic and can not be used for read regrouping")
+        }
+        tmp <- unlist(lapply(se[[names(colNamesNested)[i]]], "[[", colNamesNested[i]))
+        readInfo[[paste0(names(colNamesNested)[i], "_", colNamesNested[i])]] <- tmp
+    }
     # additional columns from colData(se)
-    for (m in setdiff(colNames, colnames(readInfo))) {
+    for (m in colNamesTopLevel) {
         if (!is.atomic(se[[m]])) {
             cli_abort("The {.var {m}} column is not atomic and can not be used for read regrouping")
         }
-        readInfo[[m]] <- rep(se[[m]], vapply(se$readInfo, nrow, 0L))
+        readInfo[[m]] <- rep(se[[m]], nReadsPerSample)
     }
     # generate read groups
     if (withinSample) {
         readGroups <- split(x = readInfo$read_id,
-                            f = apply(readInfo[, c("sample", colNames), drop = FALSE],
+                            f = apply(readInfo[, c("sample",
+                                                   paste0(names(colNamesNested),
+                                                          rep("_", length(colNamesNested)),
+                                                          colNamesNested),
+                                                   colNamesTopLevel), drop = FALSE],
                                       1, paste, collapse = "-"))
     } else {
         readGroups <- split(x = readInfo$read_id,
-                            f = apply(readInfo[, colNames, drop = FALSE],
+                            f = apply(readInfo[, c(paste0(names(colNamesNested),
+                                                          rep("_", length(colNamesNested)),
+                                                          colNamesNested),
+                                                   colNamesTopLevel), drop = FALSE],
                                       1, paste, collapse = "-"))
     }
     regroupReads(se = se, readGroups = readGroups)